+++ python/feldera/sql_context.py
@@ -20,13 +20,7 @@ from feldera._helpers import ensure_dataframe_has_columns
 from feldera.formats import JSONFormat, CSVFormat, AvroFormat
 from feldera._helpers import validate_connector_input_format
 from feldera.resources import Resources
-from enum import Enum
-
-
-class BuildMode(Enum):
-    CREATE = 1
-    GET = 2
-    GET_OR_CREATE = 3
+from feldera.enums import BuildMode, CompilationProfile
 
 
 def _table_name_from_sql(ddl: str) -> str:
@@ -47,7 +41,9 @@ class SQLContext:
     :param program_description: The description of the program. Defaults to an empty string.
     :param storage: Set `True` to use storage with this pipeline. Defaults to False.
     :param workers: The number of workers to use with this pipeline. Defaults to 8.
-    :param pipeline_resource_config: The :class:`.PipelineResourceConfig` for the pipeline. Defaults to None.
+    :param resources: The :class:`.PipelineResourceConfig` for the pipeline. Defaults to None.
+    :param compilation_profile: The compilation profile to use when compiling the program. Defaults to
+        :class:`.CompilationProfile.OPTIMIZED`.
     """
 
     def __init__(
@@ -59,7 +55,8 @@ class SQLContext:
             program_description: str = None,
             storage: bool = False,
             workers: int = 8,
-            pipeline_resource_config: Resources = None,
+            resources: Resources = None,
+            compilation_profile: CompilationProfile = CompilationProfile.OPTIMIZED
     ):
         self.build_mode: Optional[BuildMode] = None
         self.is_pipeline_running: bool = False
@@ -96,7 +93,8 @@ class SQLContext:
         self.program_description: str = program_description or ""
         self.storage: bool = storage
         self.workers: int = workers
-        self.pipeline_resource_config = pipeline_resource_config
+        self.resources: Resources = resources
+        self.compilation_profile: CompilationProfile = compilation_profile
 
     def __build_ddl(self):
         """
@@ -120,7 +118,9 @@ class SQLContext:
 
         program = Program(self.program_name, self.ddl, self.program_description)
 
-        self.client.compile_program(program)
+        self.client.compile_program(program, {
+            "profile": self.compilation_profile.value
+        })
 
         attached_cons = []
 
@@ -137,8 +137,8 @@ class SQLContext:
                 attached_cons.append(attached_con)
 
         config = { 'storage': self.storage, 'workers': self.workers }
-        if self.pipeline_resource_config:
-            config["resources"] = self.pipeline_resource_config.__dict__
+        if self.resources:
+            config["resources"] = self.resources.__dict__
 
         pipeline = Pipeline(
             self.pipeline_name,