+++ python/tests/test_program.py
@@ -6,6 +6,8 @@ import sys
 
 from tests import TEST_CLIENT
 from feldera.rest.program import Program
+from feldera.enums import CompilationProfile
+from typing import Optional
 
 NAME = str(uuid.uuid4())
 
@@ -17,7 +19,7 @@ class TestProgram(unittest.TestCase):
     handler.setLevel(logging.DEBUG)
     root.addHandler(handler)
 
-    def test_compile_program(self, name: str = NAME, delete: bool = True):
+    def test_compile_program(self, name: str = NAME, config: Optional[dict] = None, delete: bool = True):
         sql = """
     CREATE TABLE test_table (
         id INT,
@@ -26,7 +28,7 @@ class TestProgram(unittest.TestCase):
     CREATE VIEW V AS SELECT * FROM test_table;
     """
         program = Program(name, sql)
-        TEST_CLIENT.compile_program(program)
+        TEST_CLIENT.compile_program(program, config)
         assert program.version == 1
 
         if delete:
@@ -50,6 +52,13 @@ class TestProgram(unittest.TestCase):
         assert program.name == NAME
         assert program.version == 1
         self.test_delete_program(NAME)
+
+    def __test_compilation_profile(self, profile: CompilationProfile = None):
+        self.test_compile_program(str(uuid.uuid4()), config={"profile": profile.value}, delete=True)
+
+    def test_compilation_profiles(self):
+        for profile in CompilationProfile:
+            self.__test_compilation_profile(profile)
 
 
 if __name__ == '__main__':